requirements.txt
# 音高曲线比对系统依赖包
# 音频处理核心库
parselmouth>=0.4.2
numpy>=1.21.0
matplotlib>=3.5.0
scipy>=1.7.0
librosa>=0.8.1

# 音频格式处理
pydub>=0.25.1

# 机器学习和数据分析
scikit-learn>=1.0.0
dtaidistance>=2.3.4

# Web框架
flask>=2.0.0
flask-cors>=3.0.10
flask-socketio>=5.3.0
redis>=4.5.0

# TTS语音合成
edge-tts>=6.1.0     # 免费的Edge TTS (备用)
pyttsx3>=2.90       # 离线TTS (备用)

# 其他工具库
requests>=2.25.0
python-dotenv>=0.19.0

# 阿里达摩院语音处理
dashscope>=1.0.0
funasr>=1.0.0

# 文本处理和分词
jieba>=0.42.1
